AI_parking_detection/requirements.txt
# Plik: requirements.txt
#
# WAŻNE: Do instalacji tych pakietów wymagany jest zaktualizowany pip.
# Uruchom: python -m pip install --upgrade pip
# -----------------------------------------------------------------

# Główne zależności projektu
numpy
opencv-contrib-python
opencv-python
pillow

# Wsparcie dla YouTube
yt-dlp